refactor: 도메인별로 정리 및 gradle 정리
