+++ back/src/main/java/com/ms/back/member/controller/AdminController.java
@@ -1,4 +1,4 @@
-package com.ms.back.controller;
+package com.ms.back.member.controller;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
