+++ back/src/main/java/com/ms/back/member/jwt/CustomLogoutFilter.java
@@ -1,6 +1,6 @@
-package com.ms.back.jwt;
+package com.ms.back.member.jwt;
 
-import com.ms.back.repository.RefreshRepository;
+import com.ms.back.member.repository.RefreshRepository;
 import io.jsonwebtoken.ExpiredJwtException;
 import jakarta.servlet.FilterChain;
 import jakarta.servlet.ServletException;
