+++ back/src/main/java/com/ms/back/member/jwt/JWTFilter.java
@@ -1,8 +1,7 @@
-package com.ms.back.jwt;
+package com.ms.back.member.jwt;
 
-import com.ms.back.dto.CustomUserDetails;
-import com.ms.back.entity.UserEntity;
-import com.ms.back.repository.UserRepository;
+import com.ms.back.member.dto.CustomUserDetails;
+import com.ms.back.member.entity.UserEntity;
 import io.jsonwebtoken.ExpiredJwtException;
 import jakarta.servlet.FilterChain;
 import jakarta.servlet.ServletException;
