+++ back/src/main/java/com/ms/back/member/service/CustomUserDetailsService.java
@@ -1,8 +1,8 @@
-package com.ms.back.service;
+package com.ms.back.member.service;
 
-import com.ms.back.dto.CustomUserDetails;
-import com.ms.back.entity.UserEntity;
-import com.ms.back.repository.UserRepository;
+import com.ms.back.member.dto.CustomUserDetails;
+import com.ms.back.member.entity.UserEntity;
+import com.ms.back.member.repository.UserRepository;
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.security.core.userdetails.UserDetailsService;
 import org.springframework.security.core.userdetails.UsernameNotFoundException;
